+++ envision_product/tools/data_tool/__init__.py
@@ -1,3 +1,3 @@
 """RIQ Rate API package."""
-
+ 
 __version__ = "1.0.0"